cli and everything
