+++ package.json
@@ -25,6 +25,6 @@
   },
   "packageManager": "pnpm@7.18.2",
   "dependencies": {
-    "typescript": "^4.9.5"
+    "typescript": "^5.2.2"
   }
 }
