+++ packages/dynamo-store/indexer/src/index.ts
@@ -3,3 +3,4 @@ export * as AppendsEpoch from "./AppendsEpoch.js"
 export * as AppendsIndex from "./AppendsIndex.js"
 export { DynamoStoreIndexer, DynamoStoreIngester } from "./DynamoStoreIndexer.js"
 export * from "./Identifiers.js"
+export * from "./Checkpoint.js"
