+++ packages/dynamo-store/infra/lib/indexer-lambda.ts
@@ -1,5 +1,6 @@
 import { Construct } from "constructs"
 import * as cdk from "aws-cdk-lib"
+import { NodejsFunction } from "aws-cdk-lib/aws-lambda-nodejs"
 import * as lambda from "aws-cdk-lib/aws-lambda"
 import * as dynamodb from "aws-cdk-lib/aws-dynamodb"
 
@@ -12,10 +13,10 @@ export class IndexerLambda extends Construct {
   ) {
     super(scope, id)
 
-    const fn = new lambda.Function(this, "IndexerLambda", {
-      runtime: lambda.Runtime.NODEJS_18_X,
-      code: lambda.Code.fromAsset("./node_modules/@equinox-js/dynamo-store-indexer-lambda/dist"),
-      handler: "index.handler",
+    const fn = new NodejsFunction(this, "IndexerLambda", {
+      runtime: lambda.Runtime.NODEJS_16_X,
+      entry: "./node_modules/@equinox-js/dynamo-store-indexer-lambda/src/index.ts",
+      handler: "handler",
       environment: {
         TABLE_NAME: eventsTable.tableName,
         INDEX_TABLE_NAME: indexTable.tableName,
