+++ packages/dynamo-store/lambda/package.json
@@ -1,23 +1,23 @@
 {
   "name": "@equinox-js/dynamo-store-indexer-lambda",
   "scripts": {
-    "build": "tsup src/index.ts --format cjs --clean --out-dir dist"
+    "build": "tsup src/index.ts --format cjs --clean"
   },
   "devDependencies": {
-    "@equinox-js/core": "workspace:*",
-    "@equinox-js/dynamo-store": "workspace:*",
-    "@equinox-js/dynamo-store-indexer": "workspace:*",
-    "@opentelemetry/api": "^1.4.0",
     "@types/aws-lambda": "^8.10.110",
     "@types/node": "^18.11.18",
     "@vercel/ncc": "^0.36.1",
-    "tsup": "^6.5.0",
-    "typescript": "^4.7.4"
+    "tsup": "^7.2.0",
+    "typescript": "^5.2.2"
   },
   "dependencies": {
-    "@aws-sdk/client-dynamodb-streams": "^3.262.0",
     "@aws-sdk/client-dynamodb": "^3.258.0",
+    "@aws-sdk/client-dynamodb-streams": "^3.262.0",
     "@aws-sdk/smithy-client": "^3.257.0",
-    "@aws-sdk/types": "^3.257.0"
+    "@aws-sdk/types": "^3.257.0",
+    "@equinox-js/core": "workspace:*",
+    "@equinox-js/dynamo-store": "workspace:*",
+    "@equinox-js/dynamo-store-indexer": "workspace:*",
+    "@opentelemetry/api": "^1.4.0"
   }
 }
