+++ packages/dynamo-store/source/src/index.mts
@@ -1,0 +1,2 @@
+export * from './DynamoStoreSource.mjs'
+export * from './Checkpoints.js'
