+++ packages/memory-store/package.json
@@ -16,8 +16,8 @@
   "version": "1.0.0-alpha.24",
   "devDependencies": {
     "@types/node": "^18.11.18",
-    "tsup": "^6.5.0",
-    "typescript": "^4.7.4",
+    "tsup": "^7.2.0",
+    "typescript": "^5.2.2",
     "vitest": "^0.34.1"
   },
   "dependencies": {
