+++ packages/message-db/message-db-consumer/package.json
@@ -23,8 +23,8 @@
     "@types/node": "^18.11.18",
     "@types/pg": "^8.6.6",
     "p-limit": "^4.0.0",
-    "tsup": "^7.1.0",
-    "typescript": "^4.7.4",
+    "tsup": "^7.2.0",
+    "typescript": "^5.2.2",
     "vitest": "^0.34.1"
   },
   "dependencies": {
