+++ packages/message-db/message-db/package.json
@@ -20,12 +20,12 @@
   "devDependencies": {
     "@types/node": "^18.11.18",
     "@types/pg": "^8.6.6",
-    "tsup": "^6.5.0",
-    "typescript": "^4.7.4"
+    "tsup": "^7.2.0",
+    "typescript": "^5.2.2"
   },
   "dependencies": {
+    "@equinox-js/core": "workspace:*",
     "@opentelemetry/api": "^1.4.0",
-    "pg": "^8.8.0",
-    "@equinox-js/core": "workspace:*"
+    "pg": "^8.8.0"
   }
 }
