+++ packages/projection-pg/package.json
@@ -22,8 +22,8 @@
     "@types/knex": "^0.16.1",
     "@types/node": "^18.11.18",
     "@types/pg": "^8.6.6",
-    "tsup": "^6.5.0",
-    "typescript": "^4.7.4",
+    "tsup": "^7.2.0",
+    "typescript": "^5.2.2",
     "vitest": "^0.34.1"
   },
   "dependencies": {
