+++ pnpm-workspace.yaml
@@ -1,4 +1,5 @@
 packages:
-  - "packages/**"
+  - "packages/*"
+  - "packages/*/*"
   - "apps/*"
   - "docs"
